refactor(grid): Extract grid trading into a standalone grid package
Move grid trading out of the trade package into its own `grid` package,
mirroring the `dca` package: `grid.GridContext` with NewFromCfg/NewFromEnv,
and unprefixed method names (Submit/Replace/List/ListByIds/Detail/
TriggerHistory/Cancel/Suspend/Restart/OrderInfo). Types keep the Grid prefix.
